# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for the OweSome expense splitting application",
        "title": "OweSome Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "get": {
                "description": "Get list of all users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Expense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expenseShares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExpenseShare"
                    }
                },
                "group": {
                    "$ref": "#/definitions/models.Group"
                },
                "groupID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "paidBy": {
                    "$ref": "#/definitions/models.User"
                },
                "paidByID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ExpenseShare": {
            "type": "object",
            "properties": {
                "amountOwed": {
                    "type": "number"
                },
                "expense": {
                    "$ref": "#/definitions/models.Expense"
                },
                "expenseID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isPaid": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.Group": {
            "type": "object",
            "properties": {
                "adminID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Expense"
                    }
                },
                "groupAdmin": {
                    "$ref": "#/definitions/models.User"
                },
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GroupMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "profileImage": {
                    "type": "string"
                },
                "settlements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Settlement"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.GroupMember": {
            "type": "object",
            "properties": {
                "group": {
                    "$ref": "#/definitions/models.Group"
                },
                "groupID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "joinedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.Settlement": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "group": {
                    "$ref": "#/definitions/models.Group"
                },
                "groupID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isConfirmed": {
                    "type": "boolean"
                },
                "payer": {
                    "$ref": "#/definitions/models.User"
                },
                "payerID": {
                    "type": "integer"
                },
                "receiver": {
                    "$ref": "#/definitions/models.User"
                },
                "receiverID": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "adminGroups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Group"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expenseShares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExpenseShare"
                    }
                },
                "groupMemberships": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GroupMember"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "paidExpenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Expense"
                    }
                },
                "password": {
                    "type": "string"
                },
                "paymentsMade": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Settlement"
                    }
                },
                "paymentsReceived": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Settlement"
                    }
                },
                "phone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "jwt",
            "in": "cookie"
        }
    }
}